+++ sdk/search/azure-search-documents/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 11.5.0-beta.6 (Unreleased)
 
 ### Features Added
+- Added interfaces from `com.azure.core.client.traits` to `SearchIndexerClientBuilder`, `SearchIndexClientBuilder` 
+  and `SearchClientBuilder`
 
 ### Breaking Changes